Write basic tests for IClassTypes
